# lib/api-zod/src/generated/types/createUploadRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateUploadRequestContentType } from "./createUploadRequestContentType";

export interface CreateUploadRequest {
  contributorName?: string;
  contentType: CreateUploadRequestContentType;
  targetSections: string[];
  rawText: string;
}