# lib/api-zod/src/generated/types/sectionVersion.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SectionVersion {
  id: number;
  sectionId: number;
  bodyMarkdown: string;
  keyInsights: string[];
  createdAt: Date;
  createdByUploadId?: number | null;
}