# lib/api-zod/src/generated/types/sectionWithContent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SectionWithContent {
  id: number;
  slug: string;
  title: string;
  description: string;
  displayOrder: number;
  bodyMarkdown: string;
  keyInsights: string[];
  lastUpdated: Date;
}