# lib/api-zod/src/generated/types/upload.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UploadStatus } from "./uploadStatus";

export interface Upload {
  id: number;
  contributorName?: string | null;
  contentType: string;
  targetSections: string[];
  rawText: string;
  filePath?: string | null;
  status: UploadStatus;
  createdAt: Date;
  processedAt?: Date | null;
}